refactor: 상품 목록 어댑터 패키지 이동
